Spend extra time on last channel when in ADVERT_HOP mode
Hop interval is always much greater than advertising event interval
in practice, so it doesn't hurt to listen extra. The current logic
doesn't handle extending the time spent on channel 38 or 39 when a
scan request is received (since the 38 -> 39 hop uses a preset duration).
Thus, when there are lots of scan request on 38 and 39, we could
miss connect requests at the end of the 39 window.