Support CC1352 boards too
run "make PLATFORM=CC1352R1F3"

Not tested, but it **should** work on revision E silicon.
I only have a revision C CC1352 board, and it doesn't work
on it due to the updated SDK.

This currently lacks logic to configure the RF switch on
CC1352P1 launchpad boards, so this is only for CC1352R1.