Implement support for BT5 extended advertising
Adds a new -e option to the Python CLI, and parsing for extended
advertising packet headers.

When extended/auxiliary advertising mode is enabled, the sniffer
firmware will listen for ADV_EXT_IND packets on primary advertising
channels, and schedule listening on the appropriate PHY/channel to
receive the aux packet in accordance with the AuxPtr field of the
parent. The firmware will also handle connection requests made over
the secondary advertising channels.

One limitation I discovered during testing is that CMD_BLE5_GENERIC_RX
can only receive a maximum of 37 bytes unless something is overridden
to extend the receive length. The CC13x2/CC26x2 reference manual
indicates this can be overridden, but doesn't say how to override it.
I have asked TI for information on how to do this.