Improve adv hop and extended advertising coexistence
Allows a respectable best effort to capture connection establishment on
both legacy and extended advertising mechanisms simultaneously.

The key reason/issue for this change is that when a device is doing both
legacy and extended advertising, the two advertising sets do their
37/38/39 primary advertising channel hopping independently. The old
advertising channel hop logic was getting confused by the mix of legacy
and extended advertising with different hop intervals, so it kept
re-syncing again and again.

The new logic separates out extended advertising, and makes no attempt
to hop along with aux pointers on 37/38/39 (since there is no benefit:
connection establishment happens on the aux channel). It only hops along
with legacy advertisements (since connection establishment can happen on
any of the three primary channels there). The new logic also tries to
minimize the intrusiveness of confirming the sync by avoiding checks
when there is an aux packet scheduled soon.