messenger: make sure entire message is sent
It turns out that even in blocking mode, UART_write does not guarantee
the entire buffer was sent. If there is already a lot of data in the
transmit queue, it might only sent part of the supplied buffer, and
return after that portion was sent. In crowded RF environments, this
led to incomplete messages sent over UART, and was causing the Python
receiver code to error out with "SniffleHWPacketError: Incorrect length
field!"