RadioWrapper: fix channel numbering race condition
When in adv hopping mode, if we received data on 37, but the hop trigger
fired before it made it up to this interrupt, the channel number would
be mistakenly labelled as 39 instead of 37.

Note that timestamps are for the start of the packet, and each symbol
takes one microsecond, so advertisements can take hundreds of
microseconds of air time.