Fix receiving AUX_CHAIN_IND packets
I was spending too much time on the first aux channel, preventing a jump
to the next before the deadline.

Be awre that this is still not perfect, because of receive latency.
There is a 100-200 us delay from the end of packet transmission to when
my software is able to process the packet. If the next AUX_CHAIN_IND is
happening very soon after that (less than 200-ish microseconds from when
I process the last one), I will miss the next one. This usually happens
when the aux PHY is 1M or coded. When the aux PHY is 2M, the short
transmission duration usually means there's enough time to tune before
the next AUX_CHAIN_IND.