Implement instahop feature
For sniffer operation, when connection is encrypted, and hop timing can
change in encrypted control PDUs, allow just hopping to next channel
immediately when there is no more data on current channel. This feature
can be improved later by detecting changes to the connection interval,
and supporting non-zero slave connection latency, but for now this
feature works in a minimum viable form.