conf_queue: semi-reentrancy and using past events
Not a perfect fix for concurrency, but since we added the command to
manually add a channel map to the radio config queue, there are race
conditions possible in enqueing, even if unlikely. This at least makes
two radio config enqueing operations happening in different tasks
reentrantly somewhat safe.

Also, if there is a pending configuration update, this will ensure the
channel map update that was manually added won't be lost by ensuring
queue entries with instants in the past are still used. Instead, we now
assume all entries in the queue are supposed to be applied in the order
they are enqueued.