HACK: slow baud rate for less XDS110 latency
Workaround for this issue:
https://e2e.ti.com/support/wireless-connectivity/bluetooth-group/bluetooth/f/bluetooth-forum/1009230/launchxl-cc26x2r1-xds110-uart-is-bursty-and-very-high-latency-under-certain-conditions

Unfortunately, TI seems uninterested in fixing this limitation of
the XDS110 firmware. I could recommend using a different USB to UART
adapter, but that would be inconvenient for most users. The latency
spikes from fast baud rates on XDS110 are more of a problem to
operation in a connection (as master/slave) than the throughput
limitations introduced by 230400 baud.

We usually don't transmit all that much data, so the slow baud rate
is usually good enough. I deepened the queue in PacketTask to soak up
short bursts of data. This isn't ideal because Bluetooth 5 (especially
with the 2M PHY) can transmit enough to saturate this link, but
most such transfers are bursty, so the deep queue will usually take
care of that.

Revert this if using a different USB to UART adapter (such as FTDI)
or you need full throughput.