Disable RSSI filter by default
If the user wants to filter, they will filter. While the -80 default was
nice to reduce clutter on screen in my office's crowded environment, in
many other cases you want to capture everything possible. The RSSI
filter implementation in Sniffle is purely in software after the receive
too, so it made no difference to reliability of advertisement capture
either.

Note that for Launchpad hardware, while I'm setting the filter to the
minimum possible (-128), the minimum RSSI the hardware can actually
capture is around -95 to -98.