RadioWrapper: set initiator and scanner coding
For initiator and advertiser, bit 0 of the coding configuration is
ignored. Bit 1 indicates the default coding to use for requests and
responses (bit set means S=2, unset means S=8). If bit 2 is set, S=8
coding will be used regardless of default if responding to a packet with
S=8 coding. We want the bit 2 set behaviour always.