RadioTask: spend a little more time on 38
This buys a little more time to detect a CONNECT_IND on 38 in case of a
missed (too late) scan hop postponement on 37. We still hop 15 us early
to give a little margin in case of super fast adv hopping where we were
slightly late to reach 38. I don't know if there are any real world
devices that hop so quickly as to need this, but I'm leaving a little
bit of early hopping just in case.

I also parameterized the hop/retune/listen latency, and reduced how
early we listen on an aux advertising channel since it didn't need to be
that long.