RadioTask: adv hop sooner for slow hoppers
Some BLE devices (like Telink chips) hop between the advertising
channels so slowly that we can schedule a hop to 38 at 510 us after the
end of the ad on 37 (so that we can handle any SCAN_REQ for hop
postponement), and still be able to receive the advertisement on 38
because the gap between adverisements is >820 us (510 + 310 us).

I hadn't done this previously because I didn't know there are some
real devices that hop so slowly that this becomes useful.