fw: use ClockP instead of Timer
Energy efficiency aside (I don't care for this application), the
main issue with Timers is that they don't reload when you stop/start
them, and cannot be manually reloaded. They only reload when the match
value is reached. Thus, if you had a timer set up to occur in the future,
then you changed the timeout before the timer was already triggered, then
it will start counting from the wrong point and may never match until
after counter wraparound. This was breaking things like DelayStopTrigger
when trying to move the stop time to something sooner.