fw: suppress dangling pointer warning
In the TI-RTOS Semaphore_pend function, it creates a Semaphore_PendElem
on the stack and puts it in the semaphore's queue. GCC (correctly)
believes that this could result in a dangling pointer to a data
structure on the stack. However, what it does not realize is that the
function blocks until the Semaphore_post function dequeues the element
it put in the queue. As such, in reality, there will be no dangling
pointer.